+++ src/deepskylog/AstronomyLibrary/Targets/Planet.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * The target class describing a planet.
+ *
+ * PHP Version 7
+ *
+ * @category Target
+ * @author   Deepsky Developers <developers@deepskylog.be>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
+
+namespace deepskylog\AstronomyLibrary\Targets;
+
+/**
+ * The target class describing a planet.
+ *
+ * PHP Version 7
+ *
+ * @category Target
+ * @author   Deepsky Developers <developers@deepskylog.be>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
+class Planet extends Target
+{
+    /**
+     * The constructor.
+     */
+    public function __construct()
+    {
+        $this->setH0(-0.5667);
+    }
+}