+++ tests/Unit/EquatorialCoordinatesTest.php
@@ -56,39 +56,39 @@ class EquatorialCoordinatesTest extends BaseTestCase
     public function testGetSetCoordinates()
     {
         $coords = new EquatorialCoordinates(15.748, -5.42);
-        $this->assertEquals(15.748, $coords->getRA());
-        $this->assertEquals(-5.42, $coords->getDeclination());
+        $this->assertEquals(15.748, $coords->getRA()->getCoordinate());
+        $this->assertEquals(-5.42, $coords->getDeclination()->getCoordinate());
 
         $coords->setDeclination(15.2);
-        $this->assertEquals(15.2, $coords->getDeclination());
+        $this->assertEquals(15.2, $coords->getDeclination()->getCoordinate());
 
         $coords->setRA(4.2);
-        $this->assertEquals(4.2, $coords->getRA());
+        $this->assertEquals(4.2, $coords->getRA()->getCoordinate());
 
         $coords = new EquatorialCoordinates(25.748, 5.42);
-        $this->assertEquals(1.748, $coords->getRA());
+        $this->assertEquals(1.748, $coords->getRA()->getCoordinate());
 
         $coords = new EquatorialCoordinates(-1.748, -5.42);
-        $this->assertEquals(24 - 1.748, $coords->getRA());
+        $this->assertEquals(24 - 1.748, $coords->getRA()->getCoordinate());
 
         $coords = new EquatorialCoordinates(15.748, 95.42);
-        $this->assertEquals(-84.58, $coords->getDeclination());
+        $this->assertEquals(-84.58, $coords->getDeclination()->getCoordinate());
 
         $coords = new EquatorialCoordinates(15.748, -95.42);
-        $this->assertEquals(84.58, $coords->getDeclination());
+        $this->assertEquals(84.58, $coords->getDeclination()->getCoordinate());
 
         $coords = new EquatorialCoordinates(5.748, 5.42);
         $coords->setDeclination(-91.2);
-        $this->assertEquals(88.8, $coords->getDeclination());
+        $this->assertEquals(88.8, $coords->getDeclination()->getCoordinate());
 
         $coords->setDeclination(92.5);
-        $this->assertEquals(-87.5, $coords->getDeclination());
+        $this->assertEquals(-87.5, $coords->getDeclination()->getCoordinate());
 
         $coords->setRA(-1.2);
-        $this->assertEquals(22.8, $coords->getRA());
+        $this->assertEquals(22.8, $coords->getRA()->getCoordinate());
 
         $coords->setRA(24.2);
-        $this->assertEquals(0.2, $coords->getRA());
+        $this->assertEquals(0.2, $coords->getRA()->getCoordinate());
     }
 
     /**
@@ -100,8 +100,16 @@ class EquatorialCoordinatesTest extends BaseTestCase
     {
         $coords = new EquatorialCoordinates(7.7552628, 28.026183);
         $ecl = $coords->convertToEclipticalJ2000();
-        $this->assertEqualsWithDelta(113.215630, $ecl->getLongitude(), 0.00001);
-        $this->assertEqualsWithDelta(6.684170, $ecl->getLatitude(), 0.00001);
+        $this->assertEqualsWithDelta(
+            113.215630,
+            $ecl->getLongitude()->getCoordinate(),
+            0.00001
+        );
+        $this->assertEqualsWithDelta(
+            6.684170,
+            $ecl->getLatitude()->getCoordinate(),
+            0.00001
+        );
     }
 
     /**
@@ -116,8 +124,16 @@ class EquatorialCoordinatesTest extends BaseTestCase
         $location = new GeographicalCoordinates(-77.06555556, 38.92138889);
         $astrolib = new AstronomyLibrary($date, $location);
         $hor = $astrolib->equatorialToHorizontal($coords);
-        $this->assertEqualsWithDelta(68.0336, $hor->getAzimuth(), 0.0001);
-        $this->assertEqualsWithDelta(15.1249, $hor->getAltitude(), 0.0001);
+        $this->assertEqualsWithDelta(
+            68.0336,
+            $hor->getAzimuth()->getCoordinate(),
+            0.0001
+        );
+        $this->assertEqualsWithDelta(
+            15.1249,
+            $hor->getAltitude()->getCoordinate(),
+            0.0001
+        );
     }
 
     /**
@@ -129,8 +145,16 @@ class EquatorialCoordinatesTest extends BaseTestCase
     {
         $coords = new EquatorialCoordinates(23.1546225, -6.7198917);
         $gal = $coords->convertToGalactic();
-        $this->assertEqualsWithDelta(68.34653864, $gal->getLongitude(), 0.0001);
-        $this->assertEqualsWithDelta(-58.30545704, $gal->getLatitude(), 0.0001);
+        $this->assertEqualsWithDelta(
+            68.34653864,
+            $gal->getLongitude()->getCoordinate(),
+            0.0001
+        );
+        $this->assertEqualsWithDelta(
+            -58.30545704,
+            $gal->getLatitude()->getCoordinate(),
+            0.0001
+        );
     }
 
     /**